feat(settings): create PortsScreen.kt
